requirements.txt
sortedcontainers >= 2.0.4
numpy >= 1.10.4
scipy >= 1.1
pandas >= 0.17.1
simplejson >= 3.8.1
matplotlib >= 2.0.0
pyannote.core>=0.3
cvxpy >= 1.0.25
strsim == 0.0.3
tqdm >= 4.46.0
numba == 0.50.1